+++ lib/src/selectable/mp_selectable_bezier_curve_line_segment.dart
@@ -29,7 +29,7 @@ class MPSelectableBezierCurveLineSegment extends MPSelectableLineSegment {
 
     return MPNumericAux.calculateSegments(
       estimatedLength,
-      thDesiredSegmentLengthOnScreen,
+      mpDesiredSegmentLengthOnScreen,
     );
   }