+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   * New "New version available" page.
   * Created debug new version structure.
   * Improving newer version available messages.
+  * Mapiah constants should start with 'mp' prefix. Save the 'th' prefix for Therion related constants.
 
 ## 0.2.39 - 2026-03-05 - [The Seeds of Love](https://youtu.be/VAtGOESO7W8) release
 * New features: